Avoid creating unneeded inner classes
In the past our entities script made inner classes when the name of the
extension didn't match the name of the message. Even if the message
didn't have any sub messages or properties.

This is a bit of a waste since that ment unneeded code was created and
we couldn't easily update the type with the content of some other type
because the inner classes didn't match!

To avoid that we do not create inner classes anymore when it is not
technically needed.