Remove running server dependency
We no longer require a running server to generate code from all the
entities. The script will still start the same way as in reading all the
entity mapping files. After that instead of request the entity from the
server the script will just read it from the file system.

Since we are no longer requesting the server we can lose on certaining
features. For example:
    - <include... /> tag no longer gets resolved by the server. We
      accept this lose since we shouldn't use includes on the entity
      layer.